fix: Reformat and rearrange income tax-related variables
