+++ policyengine_uk/tests/policy/baseline/gov/hmrc/income_tax/bases/savings_income/taxable_savings_interest_income.yaml
@@ -1,0 +1,71 @@
+- name: No savings interest income
+  period: 2024
+  input:
+    savings_interest_income: 0
+    tax_free_savings_income: 0
+  output:
+    taxable_savings_interest_income: 0
+
+- name: Savings interest income fully covered by tax-free allowance
+  period: 2024
+  input:
+    savings_interest_income: 1000
+    tax_free_savings_income: 1000
+  output:
+    taxable_savings_interest_income: 0
+
+- name: Savings interest income partially taxable
+  period: 2024
+  input:
+    savings_interest_income: 1500
+    tax_free_savings_income: 1000
+  output:
+    taxable_savings_interest_income: 500
+
+- name: High savings interest income, standard tax-free allowance
+  period: 2024
+  input:
+    savings_interest_income: 10000
+    tax_free_savings_income: 1000
+  output:
+    taxable_savings_interest_income: 9000
+
+- name: Savings interest income, no tax-free allowance
+  period: 2024
+  input:
+    savings_interest_income: 5000
+    tax_free_savings_income: 0
+  output:
+    taxable_savings_interest_income: 5000
+
+- name: Very low positive savings interest income
+  period: 2024
+  input:
+    savings_interest_income: 1
+    tax_free_savings_income: 0
+  output:
+    taxable_savings_interest_income: 1
+
+- name: Tax-free allowance exceeds savings interest income
+  period: 2024
+  input:
+    savings_interest_income: 500
+    tax_free_savings_income: 1000
+  output:
+    taxable_savings_interest_income: 0
+
+- name: High values for both inputs
+  period: 2024
+  input:
+    savings_interest_income: 100000
+    tax_free_savings_income: 50000
+  output:
+    taxable_savings_interest_income: 50000
+
+- name: Decimal values
+  period: 2024
+  input:
+    savings_interest_income: 1234.56
+    tax_free_savings_income: 1000.00
+  output:
+    taxable_savings_interest_income: 234.56
