+++ policyengine_uk/variables/gov/hmrc/income_tax/bases/savings_income/tax_free_savings_income.py
@@ -1,0 +1,11 @@
+from policyengine_uk.model_api import *
+
+
+class tax_free_savings_income(Variable):
+    value_type = float
+    entity = Person
+    label = "Income from savings in tax-free accounts"
+    definition_period = YEAR
+    unit = GBP
+
+    adds = ["ISA_interest_income"]
