+++ policyengine_uk/variables/gov/hmrc/income_tax/bases/trading_loss.py
@@ -1,0 +1,9 @@
+from policyengine_uk.model_api import *
+
+
+class trading_loss(Variable):
+    value_type = float
+    entity = Person
+    label = "Loss from trading in the current year."
+    definition_period = YEAR
+    unit = GBP
