+++ policyengine_uk/variables/gov/hmrc/income_tax/reliefs/deficiency_relief.py
@@ -1,0 +1,13 @@
+from policyengine_uk.model_api import *
+
+"""
+The section detailing some tax reliefs applicable is section 24 of the Act, but others are described in the 2003 and 2005 Acts as deductions from the respective components.
+"""
+
+
+class deficiency_relief(Variable):
+    value_type = float
+    entity = Person
+    label = "Deficiency relief"
+    definition_period = YEAR
+    unit = GBP
